requirements.txt
# =============== WEB AUTOMATION ===============
playwright==1.40.0
pytest-playwright==0.4.3

# ================ API TESTING ================
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0

# ================ DATABASE ================
pymongo==4.5.0
dnspython==2.4.2

# ============= TESTING FRAMEWORK =============
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
allure-pytest==2.13.2
pytest-timeout==2.1.0
pytest-rerunfailures==12.0

# ================ UTILITIES ================
pandas==2.1.4
openpyxl==3.1.2
pyyaml==6.0.1
Faker==20.0.0
xlrd==2.0.1
python-dateutil==2.8.2
pytz==2023.3

# ================ LOGGING ================
colorlog==6.7.0

# =============== VALIDATION ===============
jsonschema==4.20.0

# ============ TYPE CHECKING ===========
typing-extensions==4.8.0
